Refactored code to add VesselPresenter.
Still need to refactor to add the Views.
Fixed the post build copy to work with paths containing spaces.
Fixed the default start up project to be CkerGUI.
Changed some Vessel attributes to float to try the 100 vessels file.

+++ Src/CkerModels/Vessel.cs
@@ -19,11 +19,11 @@ namespace Cker.Models
 
         public int ID { get; set; }
         public TargetType Type { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
         public float VX_0 { get; set; }
         public float VY_0 { get; set; }
-        public int StartTime { get; set; }
+        public float StartTime { get; set; }
 
         public Vessel(int id, TargetType type, int x, int y, float vx_0, float vy_0, int startTime)
         {
@@ -40,11 +40,11 @@ namespace Cker.Models
         {
             ID = Convert.ToInt32(parameters[1]);
             Type = (TargetType)Enum.Parse(typeof(TargetType), parameters[2]);
-            X = Convert.ToInt32(parameters[3]);
-            Y = Convert.ToInt32(parameters[4]);
+            X = Convert.ToSingle(parameters[3]);
+            Y = Convert.ToSingle(parameters[4]);
             VX_0 = Convert.ToSingle(parameters[5]);
             VY_0 = Convert.ToSingle(parameters[6]);
-            StartTime = Convert.ToInt32(parameters[7]);
+            StartTime = Convert.ToSingle(parameters[7]);
         }
 
         public override string ToString()
